Run dd-dotnet artifact tests on Windows
