+++ tracer/test/Datadog.Trace.Tools.dd-dotnet.ArtifactTests/Checks/ProcessBasicChecksTests.cs
@@ -4,9 +4,7 @@
 // </copyright>
 
 using System;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Datadog.Trace.TestHelpers;
@@ -18,7 +16,7 @@ using Xunit.Abstractions;
 
 using static Datadog.Trace.Tools.dd_dotnet.Checks.Resources;
 
-namespace Datadog.Trace.Tools.Runner.IntegrationTests.Checks
+namespace Datadog.Trace.Tools.dd_dotnet.ArtifactTests.Checks
 {
     public class ProcessBasicChecksTests : ConsoleTestHelper
     {
@@ -314,32 +312,6 @@ namespace Datadog.Trace.Tools.Runner.IntegrationTests.Checks
             }
 
             return false;
-        }
-
-        private async Task<string> RunTool(string arguments, params (string Key, string Value)[] environmentVariables)
-        {
-            var targetFolder = @"C:\git\dd-trace-dotnet-diag\tracer\src\Datadog.Trace.Tools.dd_dotnet\bin\Release\net8.0\win-x64\publish";
-            var executable = Path.Combine(targetFolder, RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? "dd-dotnet.exe" : "dd-dotnet");
-
-            var processStart = new ProcessStartInfo(executable, arguments)
-            {
-                RedirectStandardError = true,
-                RedirectStandardOutput = true,
-                UseShellExecute = false
-            };
-
-            foreach (var (key, value) in environmentVariables)
-            {
-                processStart.EnvironmentVariables[key] = value;
-            }
-
-            using var helper = new ProcessHelper(Process.Start(processStart));
-
-            await helper.Task;
-
-            var splitOutput = helper.StandardOutput.Split(new[] { Environment.NewLine }, StringSplitOptions.None);
-
-            return string.Join(" ", splitOutput.Select(o => o.TrimEnd()));
         }
     }
 }
